Changed Langevin integrator.
The old one had an error in it somewhere. High time steps led to a
decreased kinetic temperature (in approximately a linear fashion), even
when all potentials were turned off.

The current implementation is the original BBK implementation, not the
velocity variant. So it's now even 4th order in the positions instead of
3rd order and saves an extra forEveryParticle pass! (Though velocity
accuracy is lower.)