get timeTillZippingFromNucleation from data file
Fit still looks bad, though (with preliminary data). And it's slow as hell
trying to parse the full state. If the 'start measuring from nucleation
instead of from the quench'-criterion turns out to be good, then implement
it directly in the sampler (so loading the data can be done much quicker).