MonteCarlo: multiple pivots & cache previous V
Multiple pivot points.
It looks like it is best to decrease PIVOT_SELECTION_PROBABILITY for long
chains. Maybe even just having a fixed number (e.g. 5) of pivot points is
best? (well, unless for tiny chains -- TODO figure that out)

Jiggle moves disabled for now.

Maybe replace those jiggle moves by 'rotations'/'pivots' of the bases?
(individually, with respect to the sugars). The effect is the same as if
the bond strengths were to go to infinity (i.e. connected beads stay at the
same distance, and the bond potential is irrelevant.)