pivotMove: fixed bug + updateParameters()
The increasing of bond length was not due to numerical errors, as stated in
12b2ec19d9470fecc84a15ee0d88101bac14b4ae, was due to a bugged
implementation of the pivot moves for then-called 'pivot chains'. The
origin of the pivot point was determined by the strand at the original
position, but it should have been based on the state of the strand after
having performed all pivot moves on the 'lower' monomers and the pivot
monomer itself.

Bond lengths are now stable.