determineIdealNumberOfBoxes(): as much as possible
The fitted power law was from the time where we looped over all boxes, in that case, there was a time complexity punishment for lots of boxes.

We're only looping over occupied boxes now, so lots of boxes don't cause a time punishment (only a memory complexity punishment). For world sizes less than ~5000, memory is not really a problem.